refactor: refactor shared controllers into single file for welcome page and change password page
