+++ src/server/routes/welcome.ts
@@ -4,44 +4,20 @@ import { ResponseWithRequestState } from '@/server/models/Express';
 import { Routes } from '@/shared/model/Routes';
 import { handleAsyncErrors } from '@/server/lib/expressWrappers';
 import { logger } from '@/server/lib/logger';
-import { getBrowserNameFromUserAgent } from '@/server/lib/getBrowserName';
-import {
-  validate as validateToken,
-  change as changePassword,
-} from '@/server/lib/idapi/changePassword';
 import { renderer } from '@/server/lib/renderer';
 import { PageTitle } from '@/shared/model/PageTitle';
-import { FieldError } from '@/shared/model/ClientState';
-import { ChangePasswordErrors } from '@/shared/model/Errors';
-import { setIDAPICookies } from '@/server/lib/setIDAPICookies';
-import { trackMetric } from '@/server/lib/trackMetric';
-import { Metrics } from '@/server/models/Metrics';
 import { addReturnUrlToPath } from '@/server/lib/queryParams';
 import { consentPages } from '@/server/routes/consents';
 import { getConfiguration } from '@/server/lib/getConfiguration';
 import { resendAccountVerificationEmail } from '@/server/lib/idapi/user';
+import {
+  checkResetPasswordTokenController,
+  setPasswordTokenController,
+} from '@/server/controllers/changePassword';
 
 const router = Router();
 
 const { baseUri } = getConfiguration();
-
-const validatePasswordField = (password: string): Array<FieldError> => {
-  const errors: Array<FieldError> = [];
-
-  if (!password) {
-    errors.push({
-      field: 'password',
-      message: ChangePasswordErrors.PASSWORD_BLANK,
-    });
-  } else if (password.length < 8 || password.length > 72) {
-    errors.push({
-      field: 'password',
-      message: ChangePasswordErrors.PASSWORD_LENGTH,
-    });
-  }
-
-  return errors;
-};
 
 // resend account verification page
 router.get(
@@ -137,92 +113,25 @@ router.get(
 // welcome page, check token and display set password page
 router.get(
   `${Routes.WELCOME}${Routes.TOKEN_PARAM}`,
-  handleAsyncErrors(async (req: Request, res: ResponseWithRequestState) => {
-    let state = res.locals;
-    const { token } = req.params;
-
-    try {
-      state = deepmerge(state, {
-        pageData: {
-          browserName: getBrowserNameFromUserAgent(req.header('User-Agent')),
-          email: await validateToken(token, req.ip),
-        },
-      });
-
-      const html = renderer(`${Routes.WELCOME}/${token}`, {
-        requestState: state,
-        pageTitle: PageTitle.WELCOME,
-      });
-      return res.type('html').send(html);
-    } catch (error) {
-      logger.error(`${req.method} ${req.originalUrl}  Error`, error);
-
-      return res.redirect(`${Routes.WELCOME}${Routes.RESEND}`);
-    }
-  }),
+  checkResetPasswordTokenController(
+    Routes.WELCOME,
+    PageTitle.WELCOME,
+    Routes.WELCOME,
+    PageTitle.WELCOME,
+  ),
 );
 
 // POST form handler to set password on welcome page
 router.post(
   `${Routes.WELCOME}${Routes.TOKEN_PARAM}`,
-  handleAsyncErrors(async (req: Request, res: ResponseWithRequestState) => {
-    let state = res.locals;
-
-    const { token } = req.params;
-    const { password } = req.body;
-
-    state = deepmerge(state, {
-      pageData: {
-        browserName: getBrowserNameFromUserAgent(req.header('User-Agent')),
-      },
-    });
-
-    try {
-      const fieldErrors = validatePasswordField(password);
-
-      if (fieldErrors.length) {
-        state = deepmerge(state, {
-          pageData: {
-            email: await validateToken(token, req.ip),
-            fieldErrors,
-          },
-        });
-        const html = renderer(`${Routes.CHANGE_PASSWORD}/${token}`, {
-          requestState: state,
-          pageTitle: PageTitle.CHANGE_PASSWORD,
-        });
-        return res.status(422).type('html').send(html);
-      }
-
-      const cookies = await changePassword(password, token, req.ip);
-
-      setIDAPICookies(res, cookies);
-
-      trackMetric(Metrics.ACCOUNT_VERIFICATION_SUCCESS);
-
-      return res.redirect(
-        303,
-        addReturnUrlToPath(
-          `${Routes.CONSENTS}/${consentPages[0].page}`,
-          state.queryParams.returnUrl,
-        ),
-      );
-    } catch (error) {
-      const { message, status } = error;
-      logger.error(`${req.method} ${req.originalUrl}  Error`, error);
-
-      trackMetric(Metrics.ACCOUNT_VERIFICATION_FAILURE);
-
-      const html = renderer(`${Routes.WELCOME}/${token}`, {
-        requestState: deepmerge(state, {
-          globalMessage: {
-            error: message,
-          },
-        }),
-        pageTitle: PageTitle.WELCOME,
-      });
-      return res.status(status).type('html').send(html);
-    }
+  setPasswordTokenController(Routes.WELCOME, PageTitle.WELCOME, (res) => {
+    return res.redirect(
+      303,
+      addReturnUrlToPath(
+        `${Routes.CONSENTS}/${consentPages[0].page}`,
+        res.locals.queryParams.returnUrl,
+      ),
+    );
   }),
 );
